% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listeners.R
\docType{class}
\name{ListenerSet-class}
\alias{ListenerSet-class}
\alias{ListenerSet}
\alias{NullListenerSet}
\alias{NullListenerSet-class}
\alias{isListener,ListenerSet-method}
\alias{receiveMessage,ListenerSet-method}
\alias{notifyListeners,ListenerSet-method}
\title{Class \code{"ListenerSet"}}
\description{
This is a \dQuote{mix-in} class that adds a speaker protocol to an object,
which is complementary to the \code{\linkS4class{Listener}} protocol.  This
object maintains a list of listeners.  When the \code{notifyListeners}
method is called, it notifies each of the listeners by calling the
\code{\link{receiveMessage}} method on the listener.
}
\section{Functions}{
\itemize{
\item \code{NullListenerSet-class}: Class union containing \code{"NULL"} and \code{"ListenerSet"}
}}


\note{
The \code{notifyListeners} method uses the
\code{\link[futile.logger]{flog.logger}} protocol.  In particular, it logs
sending the message at the \dQuote{INFO} level, and the actual message sent
at the \dQuote{DEBUG} level.  In particular, setting
\code{\link[futile.logger]{flog.threshold}(DEBUG,name="Proc4")} will turn on
logging of the actual message and
\code{\link[futile.logger]{flog.threshold}(WARN,name="Proc4")} will turn off
logging of the message sent messages.

It is often useful to redirect the Proc4 logger to a log file.  In addition,
changing the logging format to JSON, will allow the message to be recovered.
Thus, try
\code{\link[futile.logger]{flog.layout}(\link[futile.logger]{layout.json},name="Proc4"}
to activate logging in JSON format.
}
\section{Extends}{


All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.  The class union \code{NullListenerSet}
is either a \code{ListenerSet} or \code{NULL}.
}

\examples{

showClass("ListenerSet")

}
\references{
\url{https://en.wikipedia.org/wiki/Observer_pattern}
}
\seealso{
\code{\link{Listener}}, \code{\link{receiveMessage}},
\code{\link{notifyListeners}}, \code{\link[futile.logger]{flog.logger}},
\code{\link[mongolite]{mongo}}, \code{\linkS4class{P4Message}}

Listener Classes.  \code{\linkS4class{CaptureListener}},
\code{\linkS4class{UpdateListener}}, \code{\linkS4class{UpsertListener}},
\code{\linkS4class{InjectionListener}}, \code{\linkS4class{TableListener}}
}
\author{
Russell Almond
}
\keyword{classes}
