% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\docType{class}
\name{MongoRec-class}
\alias{MongoRec-class}
\alias{MongoRec}
\alias{Constructor}
\alias{for}
\alias{object.}
\alias{as.jlist,MongoRec,list-method}
\alias{parse.jlist,MongoRec,list-method}
\title{Class "MongoRec".}
\usage{
MongoRec(..., m_id = NA_character_)

\S4method{as.jlist}{MongoRec,list}(obj, ml, serialize = TRUE)

\S4method{parse.jlist}{MongoRec,list}(class, rec)
}
\description{
This is a lightweight class meant to be extended.
It contains a
single field for a Mongo identifier, which can be accessed using
the \verb{m\\_id()} method.  It is meant to store something that is a
record in a Mongo collection, where \verb{\\_id} is the Mongo identifier.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.jlist}: This method actually removes the Mongo id
(\verb{_id}) as generally, that is not pass as part of an update query.

\item \code{parse.jlist}: Makes sure the \verb{_id} feild corresponds to
conventions, and interst \code{NA} if it is missing.
}}

\section{Slots}{

\describe{
\item{\code{_id}}{(character) The Mongo ID, \code{NA_character_} if not saved.}
}}

\section{Objects from the Class}{


Objects can be created by calls to the \code{MongoRec()} function.
}

\examples{

showClass("MongoRec")

}
\seealso{
\code{\link[=as.json]{as.json()}}
\code{\link[=parseObject]{parseObject()}}, \code{\link[=saveRec]{saveRec()}}, \code{\link[=getOneRec]{getOneRec()}}
}
\author{
Russell G. Almond
}
\keyword{classes}
