% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.json.R
\name{parse.jlist}
\alias{parse.jlist}
\alias{parse.jlist,ANY,list-method}
\alias{parseObject}
\title{Construct an S4 object from a list of its slot values.}
\usage{
parse.jlist(class, rec)

\S4method{parse.jlist}{ANY,list}(class, rec)

parseObject(rec, class = rec$class)
}
\arguments{
\item{class}{-- A character string defining the class of the output object.
If the list has an element named \code{class}, that will be used.}

\item{rec}{-- A list which is the output of \code{\link[jsonlite]{fromJSON}}}
}
\value{
An S4 object of type \code{class}
}
\description{

}
\details{
The \code{parse.jlist} function is a helper function designed to do any massaging
necessary to unencode the slot values before the object is produced.  The function
\code{\link{ununboxer}} undoes the effect of \code{unboxer}, and the
function \code{\link{unparseData}} udoes the effect of \code{parseData}.
}
\section{Functions}{
\itemize{
\item \code{parse.jlist}: This is the inner function for processing
the slots prior to object creation.  Generally, this is the method
that needs to be specialized.  See the \code{vignette("JSON for S4 Objects")}.

\item \code{parse.jlist,ANY,list-method}: Base case for callNextMmethod; just returns
the slot list.
}}

\examples{
\dontrun{
vignette("JSON for S4 Objects")
}
}
