% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Listeners.R
\docType{class}
\name{UpsertListener-class}
\alias{UpsertListener-class}
\alias{UpsertListener}
\alias{isListener,UpsertListener-method}
\alias{receiveMessage,UpsertListener-method}
\alias{listenerName,UpsertListener-method}
\title{Class \code{"UpsertListener"}}
\usage{
UpsertListener(
  name = "Upsert",
  sender = "sender",
  dbname = "test",
  dburi = "mongodb://localhost",
  messSet = character(),
  colname = "Messages",
  qfields = c("app", "uid"),
  ...
)
}
\description{
This listener takes messages that match its incomming set and inject them
into another Mongo database (presumably a queue for another service).  If a
matching message exists, it is replaced instead.
}
\details{
The database is a \code{\link[mongolite]{mongo}} collection identified by
\code{dburi}, \code{dbname} and \code{colname} (collection within the
database).  The \code{mess} field of the \code{\link{P4Message}} is checked
against the applicable messages in \code{messSet}.  If it is there, then the
message is saved in the collection.

Before the message is saved, the collection is checked to see if another
message exits which matches on the fields listed in \code{qfields}.  If this
is true, the message in the database is replaced.  If not, the message is
inserted.
}
\section{Extends}{


This class implements the \code{\link{Listener}} interface.

All reference classes extend and inherit methods from
\code{"\linkS4class{envRefClass}"}.
}

\examples{

\dontrun{
mess1 <- P4Message(app="default",uid="Phred",context="Down Hill",
                   sender="EABN",mess="Statistics",
                   details=list("Physics_EAP"=0.5237,"Physics_Mode"="High"))
ul <- UpsertListener(colname="Statistics",qfields=c("app","uid"),
         messSet=c("Statistics"))
receiveMessage(ul,mess1)
}

}
\references{
This is an example of the observer design pattern.
\url{https://en.wikipedia.org/wiki/Observer_pattern}.
}
\seealso{
\code{\link{Listener}}, \code{\linkS4class{P4Message}},
\code{\link{UpsertListener}}, \code{\linkS4class{UpdateListener}},
\code{\linkS4class{CaptureListener}},
\code{\linkS4class{InjectionListener}}, \code{\linkS4class{TableListener}},
\code{\link[mongolite]{mongo}}
}
\author{
Russell Almond
}
\keyword{classes}
